Change CommunityPostItem from JS to TS, refactor styles
